-OnScreenKeyboard should be usable. -Added Ability to modify strings for LCD and BIOS names. -Supports 64KB sector flash chips for persistent settings save. -Setting fan speed function will try a max of 10 times to set desired speed. -Added flash ID entries for some modchips. -Added AltBank feature.

+++ include/boot.h
@@ -454,7 +454,7 @@ typedef enum {
 u8 mbVersion;
 
 //Global to hide code when running in XBE without modchip detected.
-u8 fHasHardware;
+u16 fHasHardware;
 
 u8 currentFlashBank;
 #endif // _Boot_H_
