+++ lib/LPCMod/BootLPCMod.h
@@ -14,6 +14,6 @@
 
 void initialLPCModOSBoot(_LPCmodSettings *LPCmodSettings);
 
-u8 LPCMod_HW_rev(void);
+u16 LPCMod_HW_rev(void);
 
 #endif // _BootLPCMod_H_
